+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/action/SingleFileTemplateAction.kt
@@ -4,6 +4,7 @@ import com.github.sam0delkin.intellijpsa.services.CompletionService
 import com.github.sam0delkin.intellijpsa.settings.TemplateFormField
 import com.github.sam0delkin.intellijpsa.settings.TemplateFormFieldType
 import com.github.sam0delkin.intellijpsa.ui.components.JTextFieldCollection
+import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.ide.IdeView
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -12,6 +13,8 @@ import com.intellij.openapi.actionSystem.LangDataKeys
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.service
 import com.intellij.openapi.editor.EditorFactory
+import com.intellij.openapi.editor.event.DocumentEvent
+import com.intellij.openapi.editor.event.DocumentListener
 import com.intellij.openapi.fileEditor.OpenFileDescriptor
 import com.intellij.openapi.fileTypes.FileTypeManager
 import com.intellij.openapi.fileTypes.FileTypes
@@ -29,6 +32,9 @@ import com.intellij.ui.EditorTextField
 import com.intellij.ui.HorizontalScrollBarEditorCustomization
 import com.intellij.ui.dsl.builder.Cell
 import com.intellij.ui.dsl.builder.panel
+import com.intellij.util.textCompletion.TextCompletionValueDescriptor
+import com.intellij.util.textCompletion.TextFieldWithCompletion
+import com.intellij.util.textCompletion.ValuesCompletionProvider
 import kotlinx.serialization.json.jsonPrimitive
 import java.awt.Dimension
 import java.util.*
@@ -49,7 +55,6 @@ class SingleFileTemplateAction(
     private var loadingIcon: Cell<JLabel>? = null
     private var navigateToFile: Cell<JCheckBox>? = null
     private val formFields: HashMap<String, Cell<JComponent>> = HashMap()
-    private val richTextEditors = ArrayList<EditorTextField>()
 
     override fun actionPerformed(e: AnActionEvent) {
         val completionService = e.project?.service<CompletionService>()
@@ -101,9 +106,6 @@ class SingleFileTemplateAction(
 
                     val newFileType = FileTypeManager.getInstance()
                         .getFileTypeByFileName(templateData.get("file_name")!!.jsonPrimitive.content)
-                    richTextEditors.forEach {
-                        it.fileType = newFileType
-                    }
                     if (templateData.containsKey("file_name")) {
                         if (null !== fileNameField) {
                             fileNameField!!.component.text = templateData.get("file_name")!!.jsonPrimitive.content
@@ -180,19 +182,23 @@ class SingleFileTemplateAction(
                         }
                         formFields[field.name!!] = formField
                     } else if (field.type == TemplateFormFieldType.RichText) {
-                        val reachText = EditorTextField(
-                            EditorFactory.getInstance().createDocument(
-                                StringUtil.convertLineSeparators("")
-                            ), e.project, FileTypes.PLAIN_TEXT, false, true)
-                        reachText.addSettingsProvider { editor ->
-                            run {
-                                editor.settings.isAutoCodeFoldingEnabled = true
+                        val richText = TextFieldWithCompletion(e.project!!, ValuesCompletionProvider<String>(object: TextCompletionValueDescriptor<String> {
+                            override fun compare(o1: String?, o2: String?): Int {
+                                return o1!!.compareTo(o2!!)
                             }
-                        }
-                        reachText.preferredSize = Dimension(200, 30)
-                        reachText.maximumSize = Dimension(600, 40)
-                        richTextEditors.add(reachText)
-                        cell(reachText)
+
+                            override fun createLookupBuilder(item: String): LookupElementBuilder {
+                                return LookupElementBuilder.create(item)
+                            }
+
+                        }, if(null !== field.options) field.options!!.toList() else listOf()), "", true, true, true)
+                        richText.preferredSize = Dimension(204, 30)
+                        richText.document.addDocumentListener(object: DocumentListener {
+                            override fun documentChanged(event: DocumentEvent) {
+                                changeListener(field, event.document.text)
+                            }
+                        })
+                        cell(richText)
                     }
                 }
             }